add rope
