+++ TODO.md
@@ -2,6 +2,12 @@
   - referenced: https://github.com/0hq/WebGPT/blob/main/condensed/condensed.js
   - referenced: https://github.com/burn-rs/burn/blob/main/burn-wgpu/src/template/matmul/naive.wgsl
   - referenced: https://github.com/huggingface/candle/blob/main/candle-core/src/cpu_backend.rs
+- [ ] factor out TensorArithmetic
+  - [x] rope
+  - [ ] softmax
+  - [ ] rms_norm
+  - [ ] batch_matmul
+  - [ ] silu
 - [ ] a better gemmv
 - [ ] revise the dot product attention primitive
   - reference mlx: https://github.com/simonw/llm-mlx-llama/blob/main/llm_mlx_llama.py#L81
