+++ crabml-core/src/backends/cpu/arithmetic.rs
@@ -104,32 +104,11 @@ impl<'a, 'b> TensorArithmetics for CpuTensor<'a> {
         Ok(t)
     }
 
-    fn rope_inplace(self, pos: usize, rope_dims: usize) -> Result<Self> {
-        let mut q = self;
-        require_tensor_contiguous(&q)?;
-        require_tensor_dims(&q, &[2])?;
-
-        let n_heads = q.shape()[0];
-        let head_size = q.shape()[1];
-        let qb = q.f32_buf_mut()?;
-
-        // apply RoPE rotation for each head
-        for h in 0..n_heads {
-            for i in 0..rope_dims / 2 {
-                let theta_scale = 10000_f32.powf(-2.0 * i as f32 / head_size as f32);
-                let theta = pos as f32 * theta_scale;
-
-                let cos_theta = theta.cos();
-                let sin_theta = theta.sin();
-                let qp = &mut qb[h * head_size + i * 2..];
-                let qp0 = qp[0];
-                let qp1 = qp[1];
-                qp[0] = qp0 * cos_theta - qp1 * sin_theta;
-                qp[1] = qp0 * sin_theta + qp1 * cos_theta;
-            }
-        }
-
-        Ok(q)
+    fn rope_inplace(mut self, pos: usize, rope_dims: usize) -> Result<Self> {
+        let strider1 = self.strider().clone();
+        let buf1 = self.buf_mut();
+        primitives::rope_inplace(buf1, &strider1, pos, rope_dims)?;
+        Ok(self)
     }
 
     fn rms_norm_inplace(mut self, eps: f32) -> Result<Self> {
