+++ crabml-core/src/backends/cpu/primitives/mod.rs
@@ -1,7 +1,9 @@
 mod add_inplace;
 mod div_inplace;
 mod mul_inplace;
+mod rope_inplace;
 
 pub use add_inplace::add_inplace;
 pub use div_inplace::div_inplace;
 pub use mul_inplace::mul_inplace;
+pub use rope_inplace::rope_inplace;
